requirements.txt
adrf==0.1.3
asgiref==3.7.2
async-property==0.2.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.9
django-bootstrap-v5==1.0.11
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
fontawesomefree==6.5.1
idna==3.6
inflection==0.5.1
isort==5.12.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
soupsieve==2.5
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.0
